pyproject.toml
[project]
name = "mcp-deep-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "acres==0.5.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.11",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "brotli==1.1.0",
    "bs4>=0.0.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.2",
    "ci-info==0.3.0",
    "click==8.2.1",
    "configobj==5.0.9",
    "configparser==7.2.0",
    "cryptography==45.0.5",
    "curl-cffi==0.11.2",
    "distro==1.9.0",
    "etelemetry==0.3.1",
    "fastapi==0.116.1",
    "filelock==3.18.0",
    "frozenlist==1.6.2",
    "fsspec==2025.5.1",
    "h11==0.16.0",
    "hdbscan==0.8.40",
    "hf-xet==1.1.3",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.32.4",
    "idna==3.10",
    "ipywidgets==8.1.7",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyterlab-widgets==3.0.15",
    "looseversion==1.3.0",
    "lxml==5.4.0",
    "markdown-it-py==3.0.0",
    "markdownify==1.1.0",
    "markupsafe==3.0.2",
    "mcp[cli]==1.12.1",
    "mcpo==0.0.16",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.4.4",
    "networkx==3.5",
    "nibabel==5.3.2",
    "nipype==1.10.0",
    "numpy==2.3.0",
    "openai==1.84.0",
    "packaging==25.0",
    "pandas==2.3.0",
    "passlib==1.7.4",
    "pathlib==1.0.1",
    "pillow==11.2.1",
    "propcache==0.3.1",
    "prov==2.0.2",
    "puremagic==1.29",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydot==4.0.0",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pymupdf==1.26.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyxnat==1.6.3",
    "pyyaml==6.0.2",
    "rank-bm25==0.2.2",
    "rdflib==7.1.4",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "reportlab==4.4.1",
    "requests==2.32.3",
    "rich==14.0.0",
    "rpds-py==0.26.0",
    "safetensors==0.5.3",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "sentence-transformers==4.1.0",
    "setuptools==78.1.1",
    "shellingham==1.5.4",
    "simplejson==3.20.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sse-starlette==2.4.1",
    "starlette==0.47.2",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.1",
    "torch==2.7.1",
    "tqdm==4.67.1",
    "traits==7.0.2",
    "transformers==4.52.4",
    "typer==0.16.0",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "uv==0.8.2",
    "uvicorn==0.35.0",
    "wheel==0.45.1",
    "yarl==1.20.0",
]
